+++ src/main/scala/ai/starlake/config/Settings.scala
@@ -28,12 +28,14 @@ import ai.starlake.schema.handlers.{
   LaunchHandler,
   SimpleLauncher
 }
-import ai.starlake.schema.model.{Mode, PrivacyLevel, Sink}
+import ai.starlake.schema.model.{PrivacyLevel, Sink}
 import ai.starlake.utils.{CometObjectMapper, Utils, YamlSerializer}
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.typesafe.config.{Config, ConfigFactory, ConfigValueFactory}
 import com.typesafe.scalalogging.StrictLogging
 import org.apache.hadoop.fs.Path
+import org.apache.kafka.clients.consumer.ConsumerConfig
+import org.apache.kafka.clients.producer.ProducerConfig
 import org.apache.spark.SparkConf
 import org.apache.spark.storage.StorageLevel
 import pureconfig.ConvertHelpers._
@@ -43,7 +45,7 @@ import pureconfig.generic.auto._
 
 import java.io.ObjectStreamException
 import java.util.concurrent.TimeUnit
-import java.util.{Locale, UUID}
+import java.util.{Locale, Properties, UUID}
 import scala.collection.JavaConverters._
 import scala.concurrent.duration.FiniteDuration
 
@@ -190,22 +192,51 @@ object Settings extends StrictLogging {
 
   final case class KafkaTopicConfig(
     topicName: String,
-    topicOffsetMode: Option[Mode] = Some(Mode.STREAM), // used only for comet_offsets
     maxRead: Long = -1,
     fields: List[String] = List("key as STRING", "value as STRING"),
     partitions: Int = 1,
     replicationFactor: Short = 1,
-    writeFormat: String = "parquet",
     createOptions: Map[String, String] = Map.empty,
     accessOptions: Map[String, String] = Map.empty
-  )
+  ) {
+    def allAccessOptions(serverProperties: Map[String, String]) = {
+      serverProperties ++ accessOptions
+    }
+  }
 
   final case class KafkaConfig(
     serverOptions: Map[String, String],
     topics: Map[String, KafkaTopicConfig],
     cometOffsetsMode: Option[String] = Some("STREAM"),
     customDeserializer: Option[String]
-  )
+  ) {
+    lazy val sparkServerOptions: Map[String, String] = {
+      val ASSIGN = "assign"
+      val SUBSCRIBE_PATTERN = "subscribepattern"
+      val SUBSCRIBE = "subscribe"
+      val ignoreKafkaProperties = List(
+        ConsumerConfig.GROUP_ID_CONFIG,
+        ConsumerConfig.AUTO_OFFSET_RESET_CONFIG,
+        ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG,
+        ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG,
+        ConsumerConfig.ENABLE_AUTO_COMMIT_CONFIG,
+        ConsumerConfig.INTERCEPTOR_CLASSES_CONFIG,
+        ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG,
+        ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG,
+        ASSIGN,
+        SUBSCRIBE_PATTERN,
+        SUBSCRIBE
+      )
+      val kafkaServerProperties = new Properties()
+      serverOptions.foreach { case (k, v) =>
+        // for spark we need to prefix them with "kafka."
+        kafkaServerProperties.put(k, v)
+        if (!ignoreKafkaProperties.contains(k) && !k.startsWith("kafka."))
+          kafkaServerProperties.put(s"kafka.$k", v)
+      }
+      kafkaServerProperties.asScala.toMap
+    }
+  }
 
   case class JobScheduling(
     maxJobs: Int,
@@ -269,6 +300,7 @@ object Settings extends StrictLogging {
     analyze: Boolean,
     hive: Boolean,
     grouped: Boolean,
+    groupedMax: Int,
     mergeForceDistinct: Boolean,
     mergeOptimizePartitionWrite: Boolean,
     area: Area,
@@ -290,7 +322,9 @@ object Settings extends StrictLogging {
     defaultFileExtensions: String,
     forceFileExtensions: String,
     accessPolicies: AccessPolicies,
-    scheduling: JobScheduling
+    scheduling: JobScheduling,
+    maxParCopy: Int,
+    dsvOptions: Map[String, String]
   ) extends Serializable {
 
     val cacheStorageLevel =
@@ -341,8 +375,11 @@ object Settings extends StrictLogging {
       .fromConfig(effectiveConfig)
       .loadOrThrow[Comet]
 
+    logger.info("COMET_FS=" + System.getenv("COMET_FS"))
+    logger.info("COMET_ROOT=" + System.getenv("COMET_ROOT"))
     logger.info(YamlSerializer.serializeObject(loaded))
-    val settings = Settings(loaded, effectiveConfig.getConfig("spark"))
+    val settings =
+      Settings(loaded, effectiveConfig.getConfig("spark"), effectiveConfig.getConfig("extra"))
     val applicationConfPath = new Path(DatasetArea.metadata(settings), "application.conf")
     val result: Settings = if (settings.metadataStorageHandler.exists(applicationConfPath)) {
       val applicationConfContent = settings.metadataStorageHandler.read(applicationConfPath)
@@ -352,7 +389,12 @@ object Settings extends StrictLogging {
       val mergedSettings = ConfigSource
         .fromConfig(effectiveApplicationConfig)
         .loadOrThrow[Comet]
-      Settings(mergedSettings, effectiveApplicationConfig.getConfig("spark"))
+
+      Settings(
+        mergedSettings,
+        effectiveApplicationConfig.getConfig("spark"),
+        effectiveApplicationConfig.getConfig("extra")
+      )
     } else
       settings
     val jobConf = initSparkConfig(result)
@@ -410,6 +452,7 @@ object CometColumns {
 final case class Settings(
   comet: Settings.Comet,
   sparkConfig: Config,
+  extraConf: Config,
   jobConf: SparkConf = new SparkConf()
 ) {