+++ src/main/scala/ai/starlake/extractor/TemplateParams.scala
@@ -32,7 +32,8 @@ case class TemplateParams(
   deltaColumn: Option[String],
   dsvDelimiter: String,
   exportOutputFileBase: String,
-  scriptOutputFile: Option[File]
+  scriptOutputFile: Option[File],
+  activeEnv: Map[String, String]
 ) {
 
   val paramMap: Map[String, Any] = {
@@ -81,7 +82,7 @@ case class TemplateParams(
           "full_export"  -> fullExport
         )
       ) { case (list, deltaCol) => list :+ ("delta_column" -> deltaCol.toUpperCase) }
-      .toMap
+      .toMap ++ activeEnv
   }
 }
 
@@ -107,15 +108,17 @@ object TemplateParams {
     scriptsOutputFolder: File,
     scriptOutputPattern: Option[String],
     defaultDeltaColumn: Option[String],
-    deltaColumns: Map[String, String]
+    deltaColumns: Map[String, String],
+    activeEnv: Map[String, String]
   )(implicit settings: Settings): List[TemplateParams] =
-    domain.schemas.map(s =>
+    domain.tables.map(s =>
       fromSchema(
         domain.name,
         s,
         scriptsOutputFolder,
         scriptOutputPattern,
-        deltaColumns.get(s.name).orElse(defaultDeltaColumn)
+        deltaColumns.get(s.name).orElse(defaultDeltaColumn),
+        activeEnv
       )
     )
 
@@ -135,7 +138,8 @@ object TemplateParams {
     schema: Schema,
     scriptsOutputFolder: File,
     scriptOutputPattern: Option[String],
-    deltaColumn: Option[String]
+    deltaColumn: Option[String],
+    activeEnv: Map[String, String]
   )(implicit settings: Settings): TemplateParams = {
     val scriptOutputFileName =
       scriptOutputPattern
@@ -145,7 +149,7 @@ object TemplateParams {
             Map(
               "domain" -> domainName,
               "schema" -> schema.name
-            )
+            ) ++ activeEnv
           )
         )
         .getOrElse(s"extract_${domainName}.${schema.name}.sql")
@@ -159,12 +163,13 @@ object TemplateParams {
       tableToExport = schema.name,
       columnsToExport = schema.attributes
         .filter(_.script.isEmpty)
-        .map(col => (col.name, col.`type`, col.isIgnore(), col.privacy)),
+        .map(col => (col.name, col.`type`, col.isIgnore(), col.getPrivacy())),
       fullExport = isFullExport,
       deltaColumn = if (!isFullExport) deltaColumn else None,
       dsvDelimiter = schema.metadata.flatMap(_.separator).getOrElse(","),
       exportOutputFileBase = exportFileBase,
-      scriptOutputFile = Some(scriptsOutputFolder / scriptOutputFileName)
+      scriptOutputFile = Some(scriptsOutputFolder / scriptOutputFileName),
+      activeEnv
     )
   }
 }