+++ src/main/scala/ai/starlake/workflow/IngestionWorkflow.scala
@@ -22,14 +22,14 @@ package ai.starlake.workflow
 
 import ai.starlake.config.{DatasetArea, Settings}
 import ai.starlake.job.atlas.{AtlasConfig, AtlasJob}
-import ai.starlake.job.sink.bigquery.{BigQueryLoadConfig, BigQuerySparkJob}
-import ai.starlake.job.sink.jdbc.{ConnectionLoadConfig, ConnectionLoadJob}
-import ai.starlake.job.sink.es.{ESLoadConfig, ESLoadJob}
-import ai.starlake.job.sink.kafka.{KafkaJob, KafkaJobConfig}
 import ai.starlake.job.infer.{InferSchema, InferSchemaConfig}
 import ai.starlake.job.ingest._
 import ai.starlake.job.load.LoadStrategy
 import ai.starlake.job.metrics.{MetricsConfig, MetricsJob}
+import ai.starlake.job.sink.bigquery.{BigQueryLoadConfig, BigQuerySparkJob}
+import ai.starlake.job.sink.es.{ESLoadConfig, ESLoadJob}
+import ai.starlake.job.sink.jdbc.{ConnectionLoadConfig, ConnectionLoadJob}
+import ai.starlake.job.sink.kafka.{KafkaJob, KafkaJobConfig}
 import ai.starlake.job.transform.AutoTaskJob
 import ai.starlake.schema.generator.{Yml2DDLConfig, Yml2DDLJob}
 import ai.starlake.schema.handlers.{LaunchHandler, SchemaHandler, StorageHandler}
@@ -38,6 +38,7 @@ import ai.starlake.schema.model.Mode.{FILE, STREAM}
 import ai.starlake.schema.model._
 import ai.starlake.utils._
 import better.files.File
+import com.github.ghik.silencer.silent
 import com.google.cloud.bigquery.JobInfo.{CreateDisposition, WriteDisposition}
 import com.google.cloud.bigquery.{Schema => BQSchema}
 import com.typesafe.scalalogging.StrictLogging
@@ -47,7 +48,9 @@ import org.apache.spark.sql.types.{StructField, StructType}
 
 import java.nio.file.{FileSystems, ProviderNotFoundException}
 import java.util.Collections
+import scala.collection.GenSeq
 import scala.collection.parallel.ForkJoinTaskSupport
+import scala.concurrent.forkjoin.ForkJoinPool
 import scala.util.{Failure, Success, Try}
 
 /** The whole worklfow works as follow :
@@ -83,9 +86,15 @@ class IngestionWorkflow(
     * file. "ack" is the default ack extension searched for but you may specify a different one in
     * the domain YML file.
     */
-  def loadLanding(): Unit = {
-    logger.info("LoadLanding")
-    domains.foreach { domain =>
+  def loadLanding(config: ImportConfig): Unit = {
+    val filteredDomains = config.includes match {
+      case Nil => domains
+      case _   => domains.filter { d => config.includes.contains(d.name) }
+    }
+    logger.info(
+      s"Loading files from Landing Zone for domains : ${filteredDomains.map(_.name).mkString(",")}"
+    )
+    filteredDomains.foreach { domain =>
       val storageHandler = settings.storageHandler
       val inputDir = new Path(domain.resolveDirectory())
       if (storageHandler.exists(inputDir)) {
@@ -214,6 +223,7 @@ class IngestionWorkflow(
     *   : includes Load pending dataset of these domain only excludes : Do not load datasets of
     *   these domains if both lists are empty, all domains are included
     */
+  @silent
   def loadPending(config: WatchConfig = WatchConfig()): Boolean = {
     val includedDomains = domainsToWatch(config)
 
@@ -262,42 +272,46 @@ class IngestionWorkflow(
         logger.info(s"""Ingest resolved file : ${pendingPaths
             .map(_.getName)
             .mkString(",")} with schema ${schema.name}""")
-        val ingestingPaths = pendingPaths.map { pendingPath =>
-          val ingestingPath = new Path(DatasetArea.ingesting(domain.name), pendingPath.getName)
-          if (!storageHandler.move(pendingPath, ingestingPath)) {
-            logger.error(s"Could not move $pendingPath to $ingestingPath")
+
+        // We group by groupedMax to avoid rateLimit exceeded when the number of grouped files is too big for some cloud storage rate limitations.
+        val groupedPendingPathsIterator =
+          pendingPaths.grouped(settings.comet.groupedMax)
+        groupedPendingPathsIterator.map { pendingPaths =>
+          val ingestingPaths = pendingPaths.map { pendingPath =>
+            val ingestingPath = new Path(DatasetArea.ingesting(domain.name), pendingPath.getName)
+            if (!storageHandler.move(pendingPath, ingestingPath)) {
+              logger.error(s"Could not move $pendingPath to $ingestingPath")
+            }
+            ingestingPath
           }
-          ingestingPath
+          val jobs = if (settings.comet.grouped) {
+            JobContext(domain, schema, ingestingPaths.toList, config.options) :: Nil
+          } else {
+            // We ingest all the files but return false if one of them fails.
+            ingestingPaths.map { path =>
+              JobContext(domain, schema, path :: Nil, config.options)
+            }
+          }
+          val (parJobs, forkJoinPool) = makeParallel(jobs.toList, settings.comet.scheduling.maxJobs)
+          val res = parJobs.map { jobContext =>
+            launchHandler.ingest(
+              this,
+              jobContext.domain,
+              jobContext.schema,
+              jobContext.paths,
+              jobContext.options
+            ) match {
+              case Failure(e) =>
+                e.printStackTrace()
+                false
+              case Success(r) => true
+            }
+          }.toList
+          forkJoinPool.foreach(_.shutdown())
+          res.forall(_ == true)
         }
-        val jobs = if (settings.comet.grouped) {
-          JobContext(domain, schema, ingestingPaths.toList, config.options) :: Nil
-        } else {
-          // We ingest all the files but return false if one of them fails.
-          ingestingPaths.map { path =>
-            JobContext(domain, schema, path :: Nil, config.options)
-          }
-        }
-        val parJobs = jobs.par
-        val forkJoinPool = new java.util.concurrent.ForkJoinPool(settings.comet.scheduling.maxJobs)
-        parJobs.tasksupport = new ForkJoinTaskSupport(forkJoinPool)
-        val res = parJobs.map { jobContext =>
-          launchHandler.ingest(
-            this,
-            jobContext.domain,
-            jobContext.schema,
-            jobContext.paths,
-            jobContext.options
-          ) match {
-            case Failure(e) =>
-              e.printStackTrace()
-              false
-            case Success(r) => true
-          }
-        }.toList
-        forkJoinPool.shutdown()
-        res.forall(_ == true)
       }
-    }
+    }.flatten
     result.forall(_ == true)
   }
 
@@ -366,7 +380,7 @@ class IngestionWorkflow(
 
   private def predicate(domain: Domain, schemasName: List[String], file: Path): Boolean = {
     schemasName.exists { schemaName =>
-      val schema = domain.schemas.find(_.name.equals(schemaName))
+      val schema = domain.tables.find(_.name.equals(schemaName))
       schema.exists(_.pattern.matcher(file.getName).matches())
     }
   }
@@ -385,7 +399,7 @@ class IngestionWorkflow(
       val ingestingPaths = config.paths
       val result = for {
         domain <- domains.find(_.name == domainName)
-        schema <- domain.schemas.find(_.name == schemaName)
+        schema <- domain.tables.find(_.name == schemaName)
       } yield ingest(domain, schema, ingestingPaths, config.options)
       result match {
         case None | Some(Success(_)) => true
@@ -396,6 +410,7 @@ class IngestionWorkflow(
     }
   }
 
+  @silent
   def ingest(
     domain: Domain,
     schema: Schema,
@@ -524,12 +539,14 @@ class IngestionWorkflow(
     ingestionResult match {
       case Success(Success(jobResult)) =>
         if (settings.comet.archive) {
-          ingestingPath.foreach { ingestingPath =>
+          val (parIngests, forkJoinPool) = makeParallel(ingestingPath, settings.comet.maxParCopy)
+          parIngests.foreach { ingestingPath =>
             val archivePath =
               new Path(DatasetArea.archive(domain.name), ingestingPath.getName)
             logger.info(s"Backing up file $ingestingPath to $archivePath")
             val _ = storageHandler.move(ingestingPath, archivePath)
           }
+          forkJoinPool.foreach(_.shutdown())
         } else {
           logger.info(s"Deleting file $ingestingPath")
           ingestingPath.foreach(storageHandler.delete)
@@ -541,6 +558,22 @@ class IngestionWorkflow(
       case Failure(exception) =>
         Utils.logException(logger, exception)
         Failure(exception)
+    }
+  }
+
+  @silent
+  private def makeParallel[T](
+    collection: List[T],
+    maxPar: Int
+  ): (GenSeq[T], Option[ForkJoinPool]) = {
+    maxPar match {
+      case 1 => (collection, None)
+      case _ =>
+        val parCollection = collection.par
+        val forkJoinPool =
+          new scala.concurrent.forkjoin.ForkJoinPool(maxPar)
+        parCollection.tasksupport = new ForkJoinTaskSupport(forkJoinPool)
+        (parCollection, Some(forkJoinPool))
     }
   }
 
@@ -560,7 +593,7 @@ class IngestionWorkflow(
     Utils.logFailure(result, logger)
   }
 
-  def buildTasks(jobName: String, jobOptions: Map[String, String]): Seq[AutoTaskJob] = {
+  def buildTasks(jobName: String, configOptions: Map[String, String]): Seq[AutoTaskJob] = {
     val job = schemaHandler.jobs(jobName)
     logger.info(job.toString)
     job.tasks.map { task =>
@@ -573,9 +606,34 @@ class IngestionWorkflow(
         Views(job.views.getOrElse(Map.empty)),
         job.getEngine(),
         task,
-        jobOptions
+        configOptions,
+        task.sink
       )(settings, storageHandler, schemaHandler)
     }
+  }
+
+  def compileAutoJob(config: TransformConfig): Seq[String] = {
+    val job = schemaHandler.jobs(config.name)
+    logger.info(job.toString)
+    val result = buildTasks(config.name, config.options).map { action =>
+      val engine = action.engine
+      logger.info(s"running with -> $engine engine")
+      engine match {
+        case BQ =>
+          val (preSQL, mainSQL, postSQL) = action.buildQueryBQ()
+          mainSQL
+        case SPARK =>
+          val (preSql, mainSQL, postSql) = action.buildQuerySpark()
+          mainSQL
+        case _ =>
+          logger.error("Should never happen")
+          s"Invalid Engine $engine"
+      }
+    }
+    result.foreach { sql =>
+      logger.info(s"""START COMPILE SQL $sql END COMPILE SQL""".stripMargin)
+    }
+    result
   }
 
   /** Successively run each task of a job
@@ -588,9 +646,10 @@ class IngestionWorkflow(
     logger.info(job.toString)
     val result: Seq[Boolean] = buildTasks(config.name, config.options).map { action =>
       val engine = action.engine
-      logger.info(s"running with $engine engine")
+      logger.info(s"running with -> $engine engine")
       engine match {
         case BQ =>
+          logger.info(s"Entering $engine engine")
           val result = action.runBQ()
           val sink = action.task.sink
           logger.info(s"BQ Job succeeded. sinking data to $sink")
@@ -610,11 +669,13 @@ class IngestionWorkflow(
               logger.info(s"Spark Job succeeded. sinking data to $sinkOption")
               sinkOption match {
                 case Some(sink) => {
-                  sink.getType() match {
-                    case SinkType.ES if settings.comet.elasticsearch.active =>
+                  sink match {
+                    case _: EsSink if settings.comet.elasticsearch.active =>
                       saveToES(action)
-                    case SinkType.BQ =>
-                      val bqSink = sink.asInstanceOf[BigQuerySink]
+                    case fsSink: FsSink if !settings.comet.sinkToFile =>
+                      maybeDataFrame.exists(dataframe => action.sinkToFS(dataframe, fsSink))
+
+                    case bqSink: BigQuerySink =>
                       val source = maybeDataFrame
                         .map(df => Right(setNullableStateOfColumn(df, nullable = true)))
                         .getOrElse(Left(action.task.getTargetPath(job.getArea()).toString))
@@ -624,7 +685,7 @@ class IngestionWorkflow(
                       val config =
                         BigQueryLoadConfig(
                           source = source,
-                          outputTable = action.task.dataset,
+                          outputTable = action.task.table,
                           outputDataset = action.task.domain,
                           sourceFormat = settings.comet.defaultFormat,
                           createDisposition = createDisposition,
@@ -635,13 +696,13 @@ class IngestionWorkflow(
                           days = bqSink.days,
                           requirePartitionFilter = bqSink.requirePartitionFilter.getOrElse(false),
                           rls = action.task.rls,
-                          options = bqSink.getOptions
+                          options = bqSink.getOptions,
+                          acl = action.task.acl
                         )
                       val result = new BigQuerySparkJob(config, None).run()
                       result.isSuccess
 
-                    case SinkType.JDBC =>
-                      val jdbcSink = sink.asInstanceOf[JdbcSink]
+                    case jdbcSink: JdbcSink =>
                       val partitions = jdbcSink.partitions.getOrElse(1)
                       val batchSize = jdbcSink.batchsize.getOrElse(1000)
                       val jdbcName = jdbcSink.connection
@@ -655,7 +716,7 @@ class IngestionWorkflow(
                         jdbcName,
                         settings.comet,
                         source,
-                        outputTable = action.task.dataset,
+                        outputTable = action.task.table,
                         createDisposition = CreateDisposition.valueOf(createDisposition),
                         writeDisposition = WriteDisposition.valueOf(writeDisposition),
                         partitions = partitions,
@@ -669,12 +730,17 @@ class IngestionWorkflow(
                         case Success(_) => true
                         case Failure(e) => logger.error("JDBCLoad Failed", e); false
                       }
+                    case _: NoneSink =>
+                      maybeDataFrame.foreach { dataframe =>
+                        dataframe.write.format("console").save()
+                      }
+                      true
                     case _ =>
-                      logger.warn("No supported Sink is activated for this job")
+                      logger.warn(s"No supported Sink is activated for this job $sink")
                       true
                   }
                 }
-                case _ =>
+                case None =>
                   logger.warn("Sink is not activated for this job")
                   true
               }
@@ -693,7 +759,7 @@ class IngestionWorkflow(
   private def saveToES(action: AutoTaskJob): Boolean = {
     val targetArea = action.task.area.getOrElse(action.defaultArea)
     val targetPath =
-      new Path(DatasetArea.path(action.task.domain, targetArea.value), action.task.dataset)
+      new Path(DatasetArea.path(action.task.domain, targetArea.value), action.task.table)
     val sink: EsSink = action.task.sink
       .map(_.asInstanceOf[EsSink])
       .getOrElse(
@@ -706,7 +772,7 @@ class IngestionWorkflow(
         id = sink.id,
         format = settings.comet.defaultFormat,
         domain = action.task.domain,
-        schema = action.task.dataset,
+        schema = action.task.table,
         dataset = Some(Left(targetPath)),
         options = sink.getOptions
       )
@@ -767,7 +833,7 @@ class IngestionWorkflow(
     // Lookup for the domain given as prompt arguments, if is found then find the given schema in this domain
     val cmdArgs = for {
       domain <- schemaHandler.getDomain(cliConfig.domain)
-      schema <- domain.schemas.find(_.name == cliConfig.schema)
+      schema <- domain.tables.find(_.name == cliConfig.schema)
     } yield (domain, schema)
 
     cmdArgs match {
@@ -789,7 +855,7 @@ class IngestionWorkflow(
   def secure(config: WatchConfig): Boolean = {
     val includedDomains = domainsToWatch(config)
     val result = includedDomains.flatMap { domain =>
-      domain.schemas.map { schema =>
+      domain.tables.map { schema =>
         if (settings.comet.hive || Utils.isRunningInDatabricks()) {
           new DummyIngestionJob(
             domain,