+++ src/test/scala/ai/starlake/TestHelper.scala
@@ -20,15 +20,13 @@
 
 package ai.starlake
 
-import ai.starlake.schema.handlers.StorageHandler
-import ai.starlake.schema.model.AutoJobDesc
-import com.dimafeng.testcontainers.{ElasticsearchContainer, KafkaContainer}
 import ai.starlake.config.{DatasetArea, Settings}
 import ai.starlake.job.ingest.LoadConfig
 import ai.starlake.schema.handlers.{SchemaHandler, SimpleLauncher, StorageHandler}
-import ai.starlake.schema.model.AutoJobDesc
+import ai.starlake.schema.model.{Attribute, AutoJobDesc}
 import ai.starlake.utils.{CometObjectMapper, Utils}
-import ai.starlake.workflow.{IngestionWorkflow, WatchConfig}
+import ai.starlake.workflow.{ImportConfig, IngestionWorkflow, WatchConfig}
+import com.dimafeng.testcontainers.{ElasticsearchContainer, KafkaContainer}
 import com.fasterxml.jackson.annotation.JsonInclude.Include
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
@@ -348,7 +346,7 @@ trait TestHelper
       withSettings.deliverTestFile(sourceDatasetPathName, targetPath)
 
       val schemaHandler = new SchemaHandler(settings.storageHandler)
-      schemaHandler.checkValidity()
+      schemaHandler.fullValidation()
 
       DatasetArea.initMetadata(metadataStorageHandler)
       DatasetArea.initDomains(storageHandler, schemaHandler.domains.map(_.name))
@@ -405,7 +403,7 @@ trait TestHelper
 
       // Load landing file
       val validator = new IngestionWorkflow(storageHandler, schemaHandler, new SimpleLauncher())
-      validator.loadLanding()
+      validator.loadLanding(ImportConfig())
     }
   }
 
@@ -421,8 +419,10 @@ trait TestHelper
   // We need to start it manually because we need to access the HTTP mapped port
   // in the configuration below before any test get executed.
   lazy val kafkaContainer: KafkaContainer = {
-    val kafkaDockerTag = "5.2.1"
-    KafkaContainer.Def(kafkaDockerTag).start()
+    val kafkaDockerImage = "confluentinc/cp-kafka"
+    val kafkaDockerTag = "7.1.0"
+    val kafkaDockerImageName = DockerImageName.parse(s"$kafkaDockerImage:$kafkaDockerTag")
+    KafkaContainer.Def(kafkaDockerImageName).start()
   }
 
   lazy val esContainer: ElasticsearchContainer = {
@@ -431,7 +431,15 @@ trait TestHelper
     val esDockerImageName = DockerImageName.parse(s"$esDockerImage:$esDockerTag")
     ElasticsearchContainer.Def(esDockerImageName).start()
   }
-
+  def deepEquals(l1: List[Attribute], l2: List[Attribute]): Boolean = {
+    l1.zip(l2).foreach { case (a1, a2) =>
+      a1.name should equal(a2.name)
+      a1.`type` should equal(a2.`type`)
+      if (a1.`type` == "struct")
+        deepEquals(a1.attributes.get, a2.attributes.get)
+    }
+    true
+  }
 }
 
 object TestHelper {